// src/types/reflection/proxy/proxy-operation.ts
export enum ProxyOperation
{
    APPLY = "apply",
    CONSTRUCT = "construct",
    DEFINE_PROPERTY = "defineProperty",
    DELETE_PROPERTY = "deleteProperty",
    GET = "get",
    GET_OWN_PROPERTY_DESCRIPTOR = "getOwnPropertyDescriptor",
    GET_PROTOTYPE_OF = "getPrototypeOf",
    HAS = "has",
    IS_EXTENSIBLE = "isExtensible",
    OWN_KEYS = "ownKeys",
    PREVENT_EXTENSIONS = "preventExtensions",
    SET = "set",
    SET_PROTOTYPE_OF = "setPrototypeOf"
}